Assignment 6

+++ src/api/models/cat-model.js
@@ -12,7 +12,7 @@ const listAllCats = async () => {
 
 const findCatById = async (id) => {
   const [rows] = await promisePool.execute(
-    "SELECT cat.*, user.name AS owner_name FROM wsk_cats WHERE cat_id = ? JOIN wsk_users user ON cat.owner = user.user_id",
+    "SELECT wsk_cats.*, wsk_users.name AS owner_name FROM wsk_cats JOIN wsk_users wsk_users ON wsk_cats.owner = wsk_users.user_id WHERE wsk_cats.cat_id = ? ",
     [id]
   );
   console.log("rows", rows);
@@ -23,7 +23,7 @@ const findCatById = async (id) => {
 };
 const findCatsByUserId = async (id) => {
   const [rows] = await promisePool.execute(
-    "SELECT cat.*, user.name AS owner_name FROM wsk_cats WHERE owner = ? JOIN wsk_users user ON cat.owner = user.user_id",
+    "SELECT cat.*, user.name AS owner_name FROM wsk_cats JOIN wsk_users user ON cat.owner = user.user_id WHERE owner = ?",
     [id]
   );
   console.log("rows", rows);
@@ -39,6 +39,7 @@ const addCat = async (cat, thumb) => {
   const sql = `INSERT INTO wsk_cats (cat_name, weight, owner, filename, birthdate)
                VALUES (?, ?, ?, ?, ?)`;
   const params = [cat_name, weight, owner, filename, birthdate];
+  console.log("params", params, "sql", sql);
   const rows = await promisePool.execute(sql, params);
   console.log("rows", rows);
   if (rows[0].affectedRows === 0) {
@@ -47,11 +48,19 @@ const addCat = async (cat, thumb) => {
   return {cat_id: rows[0].insertId};
 };
 
-const modifyCat = async (cat, id) => {
-  const sql = promisePool.format(`UPDATE wsk_cats SET ? WHERE cat_id = ?`, [
-    cat,
-    id,
-  ]);
+const modifyCat = async (cat, id, user) => {
+  let sql;
+  if (user.role === "admin") {
+    sql = promisePool.format(`UPDATE wsk_cats SET ? WHERE cat_id = ?`, [
+      cat,
+      id,
+    ]);
+  } else {
+    sql = promisePool.format(
+      `UPDATE wsk_cats SET ? WHERE cat_id = ? AND owner = ?`,
+      [cat, id, user.user_id]
+    );
+  }
   const rows = await promisePool.execute(sql);
   console.log("rows", rows);
   if (rows[0].affectedRows === 0) {
@@ -60,11 +69,17 @@ const modifyCat = async (cat, id) => {
   return {message: "success"};
 };
 
-const removeCat = async (id) => {
-  const [rows] = await promisePool.execute(
-    "DELETE FROM wsk_cats WHERE cat_id = ?",
-    [id]
-  );
+const removeCat = async (id, user) => {
+  let sql;
+  if (user.role === "admin") {
+    sql = promisePool.format(`DELETE FROM wsk_cats WHERE cat_id = ?`, [id]);
+  } else {
+    sql = promisePool.format(
+      `DELETE FROM wsk_cats WHERE cat_id = ? AND owner = ?`,
+      [id, user.user_id]
+    );
+  }
+  const [rows] = await promisePool.execute(sql);
   console.log("rows", rows);
   if (rows.affectedRows === 0) {
     return false;
